UI adjustments to be mobile friendlier
Given that the marker can now be used on a mobile with the remote
feature (Android allows port forwarding!), some of the UI elements
were not really fitting well on mobile screens, including the auth
token screen and the file explorer.

This patch fixes the auth screen so the input field doesn't
overflow, and moves the favorites section above the current dir.
view for mobiles in the file explorer